fix: exporting interfaces for claim data.
